Split test_udev.py into test_udev.py and _utils.py
It needs more structure.
