+++ tests/client-dbus/tests/udev/test_udev.py
@@ -16,360 +16,30 @@ Used to test behavior of the udev device discovery mechanism.
 """
 
 # isort: STDLIB
-import base64
-import os
 import random
-import shutil
-import signal
-import string
-import subprocess
-import sys
-import time
 import unittest
 
 # isort: THIRDPARTY
-import dbus
 import psutil
-import pyudev
 
 # isort: LOCAL
-from stratisd_client_dbus import (
-    Blockdev,
-    ManagerR1,
-    MOBlockDev,
-    MOPoolR1,
-    ObjectManager,
-    PoolR1,
-    StratisdErrors,
-    blockdevs,
-    get_object,
-    pools,
-)
-from stratisd_client_dbus._constants import TOP_OBJECT
+from stratisd_client_dbus import PoolR1, get_object
 
-from ._dm import _get_stratis_devices, remove_stratis_setup
 from ._loopback import LoopBackDevices
-
-_STRATISD = os.environ["STRATISD"]
-
-_CRYPTO_LUKS_FS_TYPE = "crypto_LUKS"
-_STRATIS_FS_TYPE = "stratis"
-
-
-def random_string(length):
-    """
-    Generates a random string with the prefix 'stratis_'
-    :param length: Length of random part of string
-    :return: String
-    """
-    return "stratis_{0}".format(
-        "".join(random.choice(string.ascii_uppercase) for _ in range(length))
-    )
-
-
-def _create_pool(name, devices, *, key_description=None):
-    """
-    Creates a stratis pool. Tries three times before giving up.
-    :param name:    Name of pool
-    :param devices:  Devices to use for pool
-    :param key_description: optional key description
-    :type key_description: str or NoneType
-    :return: result of the CreatePool D-Bus method call if it succeeds
-    :rtype: bool * str * list of str
-    :raises RuntimeError: if pool is not created after three tries
-    """
-    error_reasons = []
-    for _ in range(3):
-        (result, exit_code, error_str) = ManagerR1.Methods.CreatePool(
-            get_object(TOP_OBJECT),
-            {
-                "name": name,
-                "redundancy": (True, 0),
-                "devices": devices,
-                # pylint: disable=bad-continuation
-                "key_desc": (False, "")
-                if key_description is None
-                else (True, key_description),
-            },
-        )
-        if exit_code == StratisdErrors.OK:
-            return result
-
-        error_reasons.append(error_str)
-        time.sleep(1)
-
-    raise RuntimeError(
-        "Unable to create a pool %s %s reasons: %s" % (name, devices, error_reasons)
-    )
-
-
-def _get_pools(name=None):
-    """
-    Returns a list of all pools found by GetManagedObjects, or a list
-    of pools with names matching the specified name, if passed.
-    :param name: filter for pool name
-    :type name: str or NoneType
-    :return: list of pool information found
-    :rtype: list of (str * MOPoolR1)
-    """
-    managed_objects = ObjectManager.Methods.GetManagedObjects(
-        get_object(TOP_OBJECT), {}
-    )
-
-    return [
-        (op, MOPoolR1(info))
-        for op, info in pools(props={} if name is None else {"Name": name}).search(
-            managed_objects
-        )
-    ]
-
-
-def _get_blockdevs_for_pool(pool_object_path):
-    """
-    Get a list of the blockdevs that belong to this pool.
-    :param str pool_object_path: D-Bus object path for this pool
-    :return: a list of blockdevs representing devices in the pool
-    :rtype: list of (str * MOBlockDev)
-    """
-    managed_objects = ObjectManager.Methods.GetManagedObjects(
-        get_object(TOP_OBJECT), {}
-    )
-
-    return [
-        (op, MOBlockDev(info))
-        for op, info in blockdevs(props={"Pool": pool_object_path}).search(
-            managed_objects
-        )
-    ]
-
-
-def _get_devnodes(device_object_paths):
-    """
-    Get the device nodes belonging to these object paths.
-
-    :param blockdev_object_paths: list of object paths representing blockdevs
-    :type blockdev_object_paths: list of str
-    :returns: a list of device nodes corresponding to the object paths
-    :rtype: list of str
-    """
-    return [
-        Blockdev.Properties.Devnode.Get(get_object(op)) for op in device_object_paths
-    ]
-
-
-def _settle():
-    """
-    Wait some amount and then call udevadm settle.
-    :return: None
-    """
-    time.sleep(2)
-    subprocess.check_call(["udevadm", "settle"])
-
-
-def _wait_for_udev(fs_type, expected_paths):
-    """
-    Look for devices with ID_FS_TYPE=fs_type. Check as many times as can be
-    done in 10 seconds or until the devices found are equal to the devices
-    expected. Always get the result of at least 1 enumeration.
-    :param str fs_type: the type to look for ("stratis" or "crypto_LUKS")
-    :param expected_paths: devnodes of paths that should belong to Stratis
-    :type expected_paths: list of str
-    :return: None
-    :raises RuntimeError: if unexpected device nodes are found
-    """
-    expected_devnodes = frozenset((os.path.realpath(x) for x in expected_paths))
-    found_devnodes = None
-
-    context = pyudev.Context()
-    end_time = time.time() + 10.0
-
-    while time.time() < end_time and not expected_devnodes == found_devnodes:
-        found_devnodes = frozenset(
-            [
-                x.device_node
-                for x in context.list_devices(subsystem="block", ID_FS_TYPE=fs_type)
-            ]
-        )
-        time.sleep(1)
-
-    if expected_devnodes != found_devnodes:
-        raise RuntimeError(
-            "Found unexpected devnodes: expected devnodes: %s != found_devnodes: %s"
-            % (", ".join(expected_devnodes), ", ".join(found_devnodes))
-        )
-
-
-def _processes(name):
-    """
-    Find all process matching the given name.
-    :param str name: name of process to check
-    :return: sequence of psutil.Process
-    """
-    for proc in psutil.process_iter(["name"]):
-        try:
-            if proc.name() == name:
-                yield proc
-        except psutil.NoSuchProcess:
-            pass
-
-
-def _remove_stratis_dm_devices():
-    """
-    Remove Stratis device mapper devices, fail with a runtime error if
-    some have been missed.
-    :raises RuntimeError: if some devices are remaining
-    """
-    remove_stratis_setup()
-    if _get_stratis_devices() != []:
-        raise RuntimeError("Some devices were not removed")
-
-
-class _Service:
-    """
-    Start and stop stratisd.
-    """
-
-    def start_service(self):
-        """
-        Starts the stratisd service if it is not already started. Verifies
-        that it has not exited at the time the method returns. Verifies that
-        the D-Bus service is available.
-        """
-
-        _settle()
-
-        if list(_processes("stratisd")) != []:
-            raise RuntimeError("A stratisd process is already running")
-
-        service = subprocess.Popen(
-            [_STRATISD],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            universal_newlines=True,
-        )
-
-        dbus_interface_present = False
-        limit = time.time() + 120.0
-        while (  # pylint: disable=bad-continuation
-            time.time() <= limit
-            and not dbus_interface_present
-            and service.poll() is None
-        ):
-            try:
-                get_object(TOP_OBJECT)
-                dbus_interface_present = True
-            except dbus.exceptions.DBusException:
-                time.sleep(0.5)
-
-        time.sleep(1)
-        if service.poll() is not None:
-            raise RuntimeError(
-                "Daemon unexpectedly exited with %s, stdout: %s, stderr: %s"
-                % (
-                    service.returncode,
-                    service.stdout.readlines(),
-                    service.stderr.readlines(),
-                )
-            )
-
-        if not dbus_interface_present:
-            raise RuntimeError("No D-Bus interface for stratisd found")
-
-        self._service = service  # pylint: disable=attribute-defined-outside-init
-        return self
-
-    def stop_service(self):
-        """
-        Stops the stratisd daemon previously spawned.
-        :return: a tuple of stdout and stderr
-        """
-        self._service.send_signal(signal.SIGINT)
-        output = self._service.communicate()
-        if list(_processes("stratisd")) != []:
-            raise RuntimeError("Failed to stop stratisd service")
-
-        return output
-
-
-class _KernelKey:  # pylint: disable=attribute-defined-outside-init
-    """
-    A handle for operating on keys in the kernel keyring. The specified key will
-    be available for the lifetime of the test when used with the Python with
-    keyword and will be cleaned up at the end of the scope of the with block.
-    """
-
-    def __init__(self, key_data):
-        """
-        Initialize a key with the provided key data (passphrase).
-        :param bytes key_data: The desired key contents
-        :raises RuntimeError: if the keyctl command is not found in $PATH
-                              or a keyctl command returns a non-zero exit code
-        """
-        if shutil.which("keyctl") is None:
-            raise RuntimeError("Executable keyctl was not found in $PATH")
-
-        self.key_data = key_data
-
-    def __enter__(self):
-        """
-        This method allows _KernelKey to be used with the "with" keyword.
-        :return: The key description that can be used to access the
-                 provided key data in __init__.
-        :raises subprocess.CalledProcessError:
-        """
-        with open("/dev/urandom", "rb") as urandom_f:
-            key_desc = base64.b64encode(urandom_f.read(16)).decode("utf-8")
-
-        args = ["keyctl", "get_persistent", "@s", "0"]
-        exit_values = subprocess.run(
-            args,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            universal_newlines=True,
-            check=True,
-        )
-
-        self.persistent_id = exit_values.stdout.strip()
-
-        args = ["keyctl", "add", "user", key_desc, self.key_data, self.persistent_id]
-        subprocess.run(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True)
-
-        return key_desc
-
-    def __exit__(self, exception_type, exception_value, traceback):
-        try:
-            args = ["keyctl", "clear", self.persistent_id]
-            subprocess.run(args, check=True)
-
-            args = ["keyctl", "clear", "@s"]
-            subprocess.run(args, check=True)
-        except (RuntimeError, subprocess.CalledProcessError) as rexc:
-            if exception_value is None:
-                raise rexc
-            raise rexc from exception_value
-
-
-class _ServiceContextManager:  # pylint: disable=too-few-public-methods
-    """
-    A context manager for starting and stopping the daemon.
-    """
-
-    def __init__(self):
-        self._service = _Service()
-
-    def __enter__(self):
-        self._service.start_service()
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        (_, stderrdata) = self._service.stop_service()
-
-        print("", file=sys.stdout, flush=True)
-        print(
-            "Log output from this invocation of stratisd:", file=sys.stdout, flush=True
-        )
-        print(stderrdata, file=sys.stdout, flush=True)
-
-        return False
+from ._utils import (
+    _CRYPTO_LUKS_FS_TYPE,
+    _STRATIS_FS_TYPE,
+    _create_pool,
+    _get_devnodes,
+    _get_pools,
+    _processes,
+    _KernelKey,
+    _remove_stratis_dm_devices,
+    _ServiceContextManager,
+    _settle,
+    _wait_for_udev,
+    random_string,
+)
 
 
 class UdevAdd(unittest.TestCase):
